// src/components/Header/enum.ts
export enum EMenuLabel {
  PROFILE = "Trang cá nhân",
  DEPOSIT = "Ví",
  MANAGE = "Quản lý truyện",
  AUTHOR_DASHBOARD = "Thống kê",
  AUTHOR_WRITE_STORY = "Tạo truyện mới",
  AUTHOR_WRITE_CHAPTER = "Viết chương truyện",
  AUTHOR_POSTED_STORY = "Truyện đã đăng",
}

export enum EMenuKey {
  PROFILE = "profile",
  DEPOSIT = "deposit",
  MANAGE = "manage",
}